Huge update
Done(ish)
-Intensity multiplier for player and enemy damage
-Gameover state
---Player explosion and ship removal
---Gameover splashscreen
---Automatic log writing when player dies
-Changed black to white for enemy materials to gain contrast with
background
-Rotated Drone 45 degrees so core can be seen
-Various balance updates
-Various bugfixing
WIP
-Wall NullReference
-Bolt null reference
-Cleaning up spaghetti
-Gameover splashscreen animation
-Player shutdown after gameover

+++ SpaceFun/Assets/Scripts/Score.cs
@@ -13,7 +13,7 @@ public class Score : MonoBehaviour {
 		Reset();
 	}
 		
-	public void AddPoints(int points) {
+	public void AddPoints(float points) {
 		god.score += points;
 		myText.text = "Score : "+ god.score.ToString();
 	}
